+++ modules/hanna-sprinkles/src/_/getLang.ts
@@ -1,7 +1,9 @@
-import htmlLang from '@hugsmidjan/qj/htmlLang';
 import { DEFAULT_LANG, HannaLang } from '@reykjavik/hanna-utils/i18n';
 
 const langs: Record<string, HannaLang> = { is: 'is', en: 'en', pl: 'pl' };
 
-export const getLang = (elm?: Element): HannaLang =>
-  langs[(elm ? htmlLang(elm) : htmlLang()) || DEFAULT_LANG] || DEFAULT_LANG;
+export const getLang = (elm?: Element): HannaLang => {
+  let elmLang = (elm?.closest<HTMLElement>('[lang]') || document.documentElement).lang;
+  elmLang = elmLang.slice(0, 2).toLowerCase();
+  return langs[elmLang] || DEFAULT_LANG;
+};